Split graphld CLI parser dispatch
